+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/clientadapter/BoardHandler.java
@@ -1,0 +1,5 @@
+package de.cuuky.varo.clientadapter;
+
+public class BoardHandler {
+
+}
