+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/clientadapter/tablist/TablistHandler.java
@@ -1,4 +1,4 @@
-package de.cuuky.varo.scoreboard.tablist;
+package de.cuuky.varo.clientadapter.tablist;
 
 public class TablistHandler {
