+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/data/DataManager.java
@@ -19,6 +19,7 @@ import de.cuuky.varo.bot.discord.VaroDiscordBot;
 import de.cuuky.varo.bot.discord.register.BotRegister;
 import de.cuuky.varo.bot.telegram.VaroTelegramBot;
 import de.cuuky.varo.broadcast.Broadcaster;
+import de.cuuky.varo.clientadapter.scoreboard.ScoreboardHandler;
 import de.cuuky.varo.configuration.ConfigFailureDetector;
 import de.cuuky.varo.configuration.ConfigHandler;
 import de.cuuky.varo.configuration.config.ConfigEntry;
@@ -32,7 +33,6 @@ import de.cuuky.varo.listener.PermissionSendListener;
 import de.cuuky.varo.logger.VaroLoggerManager;
 import de.cuuky.varo.mysql.MySQLClient;
 import de.cuuky.varo.report.ReportHandler;
-import de.cuuky.varo.scoreboard.scoreboard.ScoreboardHandler;
 import de.cuuky.varo.serialize.VaroSerializeHandler;
 import de.cuuky.varo.spawns.SpawnHandler;
 import de.cuuky.varo.spigot.FileDownloader;
