+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/entity/player/VaroPlayer.java
@@ -15,6 +15,7 @@ import de.cuuky.varo.alert.AlertType;
 import de.cuuky.varo.bot.BotLauncher;
 import de.cuuky.varo.bot.discord.VaroDiscordBot;
 import de.cuuky.varo.bot.discord.register.BotRegister;
+import de.cuuky.varo.clientadapter.nametag.Nametag;
 import de.cuuky.varo.configuration.config.ConfigEntry;
 import de.cuuky.varo.configuration.messages.ConfigMessages;
 import de.cuuky.varo.entity.VaroEntity;
@@ -29,7 +30,6 @@ import de.cuuky.varo.event.VaroEvent;
 import de.cuuky.varo.game.lobby.LobbyItem;
 import de.cuuky.varo.game.state.GameState;
 import de.cuuky.varo.logger.logger.EventLogger.LogType;
-import de.cuuky.varo.scoreboard.nametag.Nametag;
 import de.cuuky.varo.serialize.identifier.VaroSerializeField;
 import de.cuuky.varo.utils.JavaUtils;
 import de.cuuky.varo.vanish.Vanish;
